refactor(VirtualList): add props helper, extract type
