+++ src/components/list/VirtualList/types.ts
@@ -1,0 +1,9 @@
+import type { AnyObject, Maybe } from '@/types/Shared/Utility.types';
+
+export type TVirtualListProps = {
+  items: AnyObject[];
+  keyField?: string,
+  minItemSize?: number;
+  pageMode?: boolean;
+  getItemClass?: (item: unknown) => Maybe<string>;
+};
