+++ src/views/Character/Options/OptionsView.vue
@@ -9,7 +9,7 @@
     @list-end="nextPage"
   >
     <virtual-grouped-list
-      :list="{ items: options, keyField: DEFAULT_ENTITY_KEY_FIELD }"
+      :list="getListProps({ items: options })"
       :grid="{ flat: showRightSide }"
       :get-group="getGroupByFirstLetter"
     >
@@ -38,8 +38,9 @@
   import { usePagination } from '@/common/composition/usePagination';
   import { OptionsFilterDefaults } from '@/types/Character/Options.types';
   import VirtualGroupedList from '@/components/list/VirtualGroupedList/VirtualGroupedList.vue';
-  import { DEFAULT_ENTITY_KEY_FIELD, DEFAULT_PAGINATION_ITEMS_LIMIT } from "@/common/const";
+  import { DEFAULT_PAGINATION_ITEMS_LIMIT } from "@/common/const";
   import { getGroupByFirstLetter } from "@/common/helpers/list";
+  import { getListProps } from "@/components/list/VirtualList/helpers";
 
   type TProps = {
     inTab?: boolean,
