+++ src/views/Inventory/Armors/ArmorsView.vue
@@ -9,7 +9,7 @@
     <virtual-grouped-list
       :grid="{ flat: showRightSide }"
       :get-group="getArmorGroup"
-      :list="{ items: armors, keyField: DEFAULT_ENTITY_KEY_FIELD }"
+      :list="getListProps({ items: armors })"
     >
       <template #default="{ item: armor }">
         <armor-link
@@ -33,7 +33,7 @@
   import { ArmorsFilterDefaults } from '@/types/Inventory/Armors.types';
   import VirtualGroupedList from "@/components/list/VirtualGroupedList/VirtualGroupedList.vue";
   import type { AnyObject } from "@/types/Shared/Utility.types";
-  import { DEFAULT_ENTITY_KEY_FIELD } from "@/common/const";
+  import { getListProps } from "@/components/list/VirtualList/helpers";
 
   const route = useRoute();
   const router = useRouter();
